chore: convert back to JSX
