+++ package.json
@@ -107,6 +107,7 @@
     "@testing-library/react": "12.1.5",
     "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "^13.2.1",
+    "@types/react-instantsearch-dom": "^6.12.7",
     "axios": "^0.27.2",
     "axios-mock-adapter": "1.22.0",
     "eslint-import-resolver-webpack": "^0.13.8",
